【要確認】Doubleに変換したら動いた

+++ erp/src/main/java/com/erp/main/domain/objects/entity/fields/QuotationEntityFileds.java
@@ -91,13 +91,13 @@ public class QuotationEntityFileds extends BaseEntityFields {
 	 */
 	@NotNull
 	@Column(name="TAX")
-	private Long tax;
+	private double tax;
 	
 	/**
 	 * 合計金額
 	 */
 	@NotNull
 	@Column(name="TOTAL")
-	private Long total;
+	private double total;
 	
 }
